+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/auth/ResetPasswordRequest.java
@@ -1,9 +1,13 @@
 package gov.healthit.chpl.domain.auth;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
+@Builder
+@AllArgsConstructor
 @NoArgsConstructor
 public class ResetPasswordRequest {